requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
joblib==1.3.2
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
patsy==0.5.3
psycopg==3.1.12
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
scikit-learn==1.3.2
scipy==1.11.3
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
statsmodels==0.14.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
